refactor: centralize api error handlers in controller
